Reorganized the Screen Well's rows to be in the correct order and constrained against the correct rows and/or well walls.  Decreased Slow Inverval for faster control repetition on hold.  Adjusted Game Well not to draw empty Traffic Light Bulbs which allows Traffic Lights to render right next to the Well's walls without erroring.  Also adjusted the on-screen controls' and Well gestures' behavior to redraw the Game Well to the Screen Well after manipulating a Traffic Light rather than only when the Game Timer fires.  This allows for a much faster, much more responsive user experience.
The Game Well and Game Repeater need to be refactored under the context of a Game type with Play, Pause, and Stop semantics, and Traffic Light Drop Down on Timer.  Also, the Game Well need to detect collision ahead of time and prevent Traffic Light manipulations that would/could render out of its bounds, and pieces that bottom out should be added to the Game Well's contents.